* [apiclient] reorganize stub files
