+++ apiclient/harvester_api/models/virtualmachines.pyi
@@ -1,52 +1,7 @@
-from typing import Type, TypeVar, TypeAlias, Final, Optional, NoReturn, ClassVar, Iterable, List
+from __future__ import annotations
+from typing import ClassVar, Optional, List, NoReturn, TypeAlias, Type
 
-RS_T = TypeVar("RS_T", bound="RestoreSpec")
-VS_T = TypeVar("VS_T", bound="VolumeSpec")
-VM_T = TypeVar("VM_T", bound="VMSpec")
-SS_T = TypeVar("SS_T", bound="BaseSettingSpec")
 United_S: TypeAlias = str
-
-MGMT_NETID: Final[object]
-DEFAULT_STORAGE_CLS: Final[str]
-
-
-class RestoreSpec:
-    """
-    """
-    new_vm: str
-    vm_name: str
-    namespace: str
-    delete_volumes: bool
-
-    def __init__(
-        self,
-        new_vm: str,
-        vm_name: Optional[str] = ...,
-        namespace: Optional[str] = ...,
-        delete_volumes: Optional[bool] = ...
-    ) -> NoReturn:
-        """
-        """
-    def to_dict(
-        self,
-        name: str,
-        namespace: str,
-        existing_vm: str
-    ) -> dict:
-        """
-        """
-    @classmethod
-    def for_new(
-        cls: Type[RS_T],
-        vm_name: str,
-        namespace: Optional[str] = ...
-    ) -> RS_T:
-        """
-        """
-    @classmethod
-    def for_existing(cls: Type[RS_T], delete_volumes: Optional[bool] = ...) -> RS_T:
-        """
-        """
 
 
 class VMSpec:
@@ -200,100 +155,6 @@ class VMSpec:
         """
         """
     @classmethod
-    def from_dict(cls: Type[VM_T], data: dict) -> VM_T:
-        """
-        """
-
-
-class VolumeSpec:
-    """
-    """
-
-    volume_mode: ClassVar[str]
-
-    access_modes: Iterable[str]
-    size: int | float | United_S
-    storage_cls: str
-    description: str
-    annotations: dict
-
-    def __init__(
-        self,
-        size: int | float | United_S,
-        storage_cls: Optional[str] = ...,
-        description: Optional[str] = ...,
-        annotations: Optional[dict] = ...
-    ) -> NoReturn:
-        """
-        """
-    def to_dict(
-        self,
-        name: str,
-        namespace: str,
-        image_id: Optional[str] = ...
-    ) -> dict:
-        """
-        """
-    @classmethod
-    def from_dict(cls: Type[VS_T], data: dict) -> VS_T:
-        """
-        """
-
-
-class BaseSettingSpec:
-    """
-    """
-
-    data: dict
-
-    def __init__(self, data: dict) -> NoReturn:
-        """
-        """
-    def to_dict(self) -> dict:
-        """
-        """
-    @classmethod
-    def from_dict(cls: Type[SS_T], data: dict) -> SS_T:
-        """
-        """
-
-
-class BackupTargetSpec(BaseSettingSpec):
-    """
-    """
-
-    value: Optional[dict]
-
-    def __init__(self, value: Optional[dict] = ...) -> NoReturn:
-        """
-        """
-    @property
-    def type(self) -> str | None:
-        """
-        """
-    def clear(self) -> NoReturn:
-        """
-        """
-    def to_dict(self) -> dict:
-        """
-        """
-    @classmethod
-    def from_dict(cls: Type[SS_T], data: dict) -> SS_T:
-        """
-        """
-    @classmethod
-    def S3(
-        cls: Type[SS_T],
-        bucket: str,
-        region: str,
-        access_id: str,
-        access_secret,
-        endpoint: str = ...,
-        virtual_hosted: Optional[bool] = ...
-    ) -> SS_T:
-        """
-        """
-    @classmethod
-    def NFS(cls: Type[SS_T], endpoint: str) -> SS_T:
+    def from_dict(cls, data: dict) -> Type[VMSpec]:
         """
         """
